Fleshing out systems

+++ Assets/Scripts/Interfaces/IHasHealth.cs
@@ -2,9 +2,9 @@ using System;
 using UnityEngine;
 
 public interface IHasHealth {
-    int HealthPoints { get; }
+    float HealthPoints { get; }
     int Armor { get; }
-    void TakeDamage(int damage);
+    void TakeDamage(float damage);
     void Kill();
     bool IsDead { get; }
     event DeathHandler<Transform> OnDeath;
